feat(css/compat): handle rgb/rgba/hsl/hsla
