+++ crates/swc_css_compat/src/compiler/color_alpha_parameter.rs
@@ -1,7 +1,7 @@
 use swc_atoms::js_word;
-use swc_css_ast::{AbsoluteColorBase, AlphaValue, ComponentValue, Number, Percentage};
+use swc_css_ast::{AbsoluteColorBase, ComponentValue};
 
-use crate::compiler::{utils::round_alpha, Compiler};
+use crate::compiler::Compiler;
 
 impl Compiler {
     pub(crate) fn process_color_alpha_parameter(&mut self, n: &mut AbsoluteColorBase) {
@@ -40,34 +40,6 @@ impl Compiler {
                         function.name.raw = None;
                     }
                     _ => {}
-                }
-            }
-        }
-    }
-
-    pub(crate) fn process_float_values_in_parameters(&mut self, n: &mut AbsoluteColorBase) {
-        if let AbsoluteColorBase::Function(function) = n {
-            if let Some(ComponentValue::AlphaValue(box alpha_value)) = function.value.last_mut() {
-                let name = function.name.value.to_ascii_lowercase();
-
-                if !matches!(
-                    name,
-                    js_word!("rgb") | js_word!("rgba") | js_word!("hsl") | js_word!("hsla")
-                ) {
-                    return;
-                }
-
-                if let AlphaValue::Percentage(Percentage {
-                    span,
-                    value: Number { value: a, .. },
-                    ..
-                }) = alpha_value
-                {
-                    *alpha_value = AlphaValue::Number(Number {
-                        span: *span,
-                        value: round_alpha(*a / 100.0),
-                        raw: None,
-                    });
                 }
             }
         }
