+++ crates/swc_css_compat/src/feature.rs
@@ -7,7 +7,7 @@ bitflags! {
         const MEDIA_QUERY_RANGES = 1 << 2;
         const COLOR_HEX_ALPHA = 1 << 3;
         const COLOR_ALPHA_PARAMETER = 1 << 4;
-        const COLOR_FLOAT_VALUES_IN_PARAMETERS = 1 << 5;
+        const COLOR_PERCENTAGE_IN_PARAMETERS = 1 << 5;
         const COLOR_SPACE_SEPARATED_PARAMETERS = 1 << 6;
     }
 }
