+++ crates/swc_css_compat/tests/color-legacy/input.expect.css
@@ -15,9 +15,9 @@
   color: rgba(178, 34, 34, CaLC(1 / 2));
 }
 .test-rgb-percentages {
-  color: rgb(70%, 13.5%, 13.5%);
-  color: rgba(70%, 13.5%, 13.5%, 1);
-  color: rgba(70%, 13.5%, 13.5%, 0.5);
+  color: rgb(70, 13.5, 13.5);
+  color: rgba(70, 13.5, 13.5, 1);
+  color: rgba(70, 13.5, 13.5, 0.5);
 }
 .test-hsl {
   color: hsl(0, 0%, 100%);
@@ -25,8 +25,8 @@
   color: hsl(120, 100%, 50%);
   color: hsla(120, 100%, 50%, 1);
   color: hsla(120, 100%, 50%, .5);
-  color: hsla(120, 100%, 50%, 1);
-  color: hsla(120, 100%, 50%, 0.5);
+  color: hsla(120, 100%, 50%, 100%);
+  color: hsla(120, 100%, 50%, 50%);
   color: hsl(120deg, 100%, 50%);
   color: hsl(0.5turn, 100%, 50%);
   color: hsl(200grad, 100%, 50%);
@@ -38,8 +38,8 @@
   color: hsl(120, 100%, 50%);
   color: hsla(120, 100%, 50%, 1);
   color: hsla(120, 100%, 50%, .5);
-  color: hsla(120, 100%, 50%, 1);
-  color: hsla(120, 100%, 50%, 0.5);
+  color: hsla(120, 100%, 50%, 100%);
+  color: hsla(120, 100%, 50%, 50%);
   color: hsla(120, 100%, 50%, var(--alpha-50));
   color: hsla(120, 100%, 50%, calc(1 / 2));
 }
@@ -55,26 +55,26 @@
   color: rgb(; );
 }
 .rgb-with-alpha {
-  color: rgba(70%, 13.5%, 13.5%, 0.5);
+  color: rgba(70, 13.5, 13.5, 0.5);
   color: rgba(0, 0, 0, 0);
 }
 .hsl-with-alpha {
-  color: hsla(120, 100%, 50%, 0.5);
+  color: hsla(120, 100%, 50%, 50%);
   color: hsla(120, 100%, 50%, 0);
 }
 .test-legacy-notation-with-modern-components--rgb {
-  color: rgb(50%, 34, 34);
-  color: rgb(178, 10%, 34);
-  color: rgb(178, 34, 10%);
+  color: rgb(50, 34, 34);
+  color: rgb(178, 10, 34);
+  color: rgb(178, 34, 10);
 }
 .test-legacy-notation-with-modern-components--rgba {
-  color: rgba(50%, 34, 34, 1);
-  color: rgba(178, 10%, 34, 1);
-  color: rgba(178, 34, 10%, 1);
+  color: rgba(50, 34, 34, 1);
+  color: rgba(178, 10, 34, 1);
+  color: rgba(178, 34, 10, 1);
   color: rgba(178, 34, 34, 1);
 }
 .test-legacy-notation-with-modern-components--hsla {
-  color: hsla(120, 100%, 50%, 1);
+  color: hsla(120, 100%, 50%, 100%);
 }
 .test-ignore {
   color: rgba(0, 0, 0, 0.1);
