+++ crates/swc_css_compat/tests/fixture.rs
@@ -159,7 +159,7 @@ fn test_color_space_separated_function_notation(input: PathBuf) {
 
         ss.visit_mut_with(&mut Compiler::new(Config {
             process: Features::COLOR_SPACE_SEPARATED_PARAMETERS
-                | Features::COLOR_FLOAT_VALUES_IN_PARAMETERS
+                | Features::COLOR_PERCENTAGE_IN_PARAMETERS
                 | Features::COLOR_ALPHA_PARAMETER,
         }));
